feat: boilerplate upgrade
