+++ blocks/carousel/carousel.js
@@ -13,9 +13,7 @@
 /* global window fetch document */
 /* eslint-disable import/named, import/extensions */
 
-import {
-  createTag,
-} from '../../scripts/scripts.js';
+import { createTag } from '../../scripts/utils.js';
 
 function masonrize($cells, $masonry, force) {
   const colWidth = 264;
